IN-111 fix: button is enabled when a new avatar is uploaded, revert previous NavLink logic
