+++ src/common/components/sideBar/navLink/NavLink.tsx
@@ -1,6 +1,5 @@
 import React, { ComponentPropsWithoutRef, ElementType } from 'react'
 
-import { getBasePath } from '@/common/components/sideBar/navLink/helpers'
 import { cn } from '@/common/utils/cn'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
@@ -36,20 +35,14 @@ export const NavLink = React.forwardRef<HTMLElement, Props<ElementType>>(
     const Component = as || Link
     const { pathname } = useRouter()
 
-    if (!href) {
-      return
-    }
-
-    const basePath = getBasePath(pathname)
-    const baseHref = getBasePath(href)
-
     let isActive = false
 
     if (Component === Link) {
       if (href === '/') {
         isActive = pathname === href
       } else {
-        isActive = basePath === baseHref
+        // isActive = basePath === baseHref
+        isActive = pathname.startsWith(href)
       }
     } else {
       isActive = !!iconTrigger
