+++ src/features/profile/settings/generalInfo/model/useFillGeneralInfo.ts
@@ -1,13 +1,12 @@
-import { useState } from 'react'
 import { useForm } from 'react-hook-form'
 
 import { useTranslation } from '@/common'
 import { responseErrorHandler } from '@/common/utils/responseErrorHandler'
 import AuthStore from '@/features/auth/model/authStore'
-import { PhotoResult } from '@/features/profile/settings/avatarDialog/model'
 import { FormData, useCalendar } from '@/features/profile/settings/generalInfo'
 import { generalInfoSchema } from '@/features/profile/settings/generalInfo/model/generalInfoSchema'
 import ProfileStore from '@/features/profile/settings/generalInfo/model/profileStore'
+import { useAvatarUpload } from '@/features/profile/settings/generalInfo/model/useAvatarUpload'
 import { useFetchLocations } from '@/features/profile/settings/generalInfo/model/useFetchLocations'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { format } from 'date-fns'
@@ -16,15 +15,6 @@ import { ru } from 'date-fns/locale'
 export const useFillGeneralInfo = () => {
   const profile = AuthStore.profile
   const userProfile = ProfileStore.userProfile
-  const currentPhoto = ProfileStore.userProfile?.avatars[0]?.url
-  const [photoObj, setPhotoObj] = useState<PhotoResult>({
-    photo: currentPhoto || null,
-    photoForServer: null,
-  })
-
-  const onModalPhotoSave = (photo: PhotoResult) => {
-    setPhotoObj(photo)
-  }
 
   const formattedDate =
     userProfile?.dateOfBirth &&
@@ -53,10 +43,14 @@ export const useFillGeneralInfo = () => {
     useCalendar(setValue, 'dateOfBirth')
   const { cities, countriesData, countryValue } = useFetchLocations(control)
   const { t } = useTranslation()
+  const { currentPhoto, onModalPhotoSave, photoChanged, photoObj } =
+    useAvatarUpload()
 
   const onSubmit = async (data: FormData) => {
     try {
-      await ProfileStore.updateProfile(data)
+      if (isDirty) {
+        await ProfileStore.updateProfile(data)
+      }
       if (currentPhoto !== photoObj.photo) {
         await ProfileStore.uploadAvatar(photoObj)
       }
@@ -79,6 +73,7 @@ export const useFillGeneralInfo = () => {
     onModalPhotoSave,
     onSelectDate,
     onSubmit,
+    photoChanged,
     photoObj,
     t,
     toggleCalendar,
